requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
autopep8==1.6.0
cffi==1.16.0
distlib==0.3.7
Django==3.2.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
filelock==3.13.1
Pillow==9.5.0
platformdirs==3.11.0
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
virtualenv==20.24.6
